requirements.txt
fastapi==0.108.0
uvicorn==0.30.0
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==2.0.2
paddlepaddle==2.5.2
paddleocr==2.7.0.3
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
